flaten ListItem and make own component for RSVP
