+++ client/src/components/List/index.jsx
@@ -3,7 +3,7 @@ import React, { Component } from 'react';
 import { withRouter } from 'react-router-dom';
 import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
-import ListItem from './ListItem';
+import ListItem from '../ListItem';
 import { addEvents } from '../actions/events';
 
 require('./styles.css');
