+++ client/src/components/ListItem/index.jsx
@@ -5,8 +5,9 @@ import axios from 'axios';
 import { deleteEvent, addAttendee, removeAttendee } from '../actions/events';
 import { withRouter } from 'react-router-dom';
 import { connect } from 'react-redux';
-// MAKE SURE ATTENDANCE STATUS PERSISTS
-const ListItem = ({ event, user, deleteEvent, addAttendee, removeAttendee }) => {
+import RSVP from '../RSVP';
+
+const ListItem = ({ event, user, deleteEvent }) => {
   function handleDelete() {
     axios.delete(`/api/event/${event._id}`)
       .then(() => {
@@ -17,50 +18,22 @@ const ListItem = ({ event, user, deleteEvent, addAttendee, removeAttendee }) => 
       });
   }
 
-  function handleJoin() {
-    axios.post(`/api/event/${event._id}/attendee`, {
-      _id: user._id,
-    })
-      .then(() => {
-        addAttendee(event._id, user._id, user.name);
-      })
-      .catch((err) => {
-        console.error('error joining event', err);
-      });
-  }
-
-  function handleLeave() {
-    axios.delete(`/api/event/${event._id}/attendee/`, {
-      _id: user._id,
-    })
-      .then(() => {
-        removeAttendee(event._id, user._id);
-      })
-      .catch((err) => {
-        console.error('error leaving event', err);
-      });
-  }
-
   return (
     <li key={event._id} className="collection-item">
       <span>
         <Link to={`/event/${event._id}`}>{event.title}</Link>
-        { event.creator !== user._id && (
-            !event.attendees.find(attendee => attendee._id === user._id)
-            ? <i className="material-icons small right" onClick={() => { handleJoin(); }}>add</i>
-            : <i className="material-icons small right" onClick={() => { handleLeave(); }}>report_problem</i>
-          )}
+        {event.creator !== user._id && <RSVP event={event}/> }
         { event.creator === user._id &&
           <span>
             <i className="material-icons small right">
               <Link to={`/edit/${event._id}`}>mode_edit</Link>
             </i>
-            <i className="material-icons small right" onClick={() => { handleDelete(event._id); }}>
-              delete
-            </i>
+            <a onClick={() => { handleDelete(event._id); }}>
+              <i className="material-icons small right">delete</i>
+            </a>
           </span>
         }
-        </span>
+      </span>
     </li>
   );
 };
@@ -74,8 +47,6 @@ ListItem.propTypes = {
   }).isRequired,
   event: PropTypes.object.isRequired,
   deleteEvent: PropTypes.func.isRequired,
-  addAttendee: PropTypes.func.isRequired,
-  removeAttendee: PropTypes.func.isRequired,
 };
 
 const mapStateToProps = ({ user, events }) => ({ user, events });
@@ -83,12 +54,6 @@ const mapStateToProps = ({ user, events }) => ({ user, events });
 const mapDispatchToProps = dispatch => ({
   deleteEvent: (_id) => {
     dispatch(deleteEvent(_id));
-  },
-  addAttendee: (event_id, user_id, userName) => {
-    dispatch(addAttendee(event_id, user_id, userName));
-  },
-  removeAttendee: (event_id, user_id) => {
-    dispatch(removeAttendee(event_id, user_id));
   },
 });
